Refactored slot interaction
Had bug where depending on order of execution an item could be removed from a clicked slot before passing the item reference to the target slot resulting in an error.
I've changed it so the target slot does both the adding and removing of the item. Before, both the target and clicked slot would listen for the same event and and add / remove the item independently.